Changed attribute infectionPressure to float (from long)

+++ src/main/java/com/example/eksamensprojekt_24_timers/model/Parish.java
@@ -19,7 +19,7 @@ public class Parish { //Sogn
     @JoinColumn(name = "commune_id")
     private Commune commune; //Kommune
 
-    private long infectionPressure; //Smittetryk
+    private float infectionPressure; //Smittetryk
 
     private LocalDate shutDownStartDate; //Start på nedlukning
